requirements.txt
streamlit>=1.35.0
spotipy>=2.23.0
pandas
scikit-learn
python-dotenv
